Fixed the issue that Jeff Julian found with too many database calls when loading comments for an individual blog post.

+++ branches/Release1.9/SubtextSolution/Subtext.Framework/ImportExport/ObjectHydrator.cs
@@ -46,7 +46,7 @@ namespace Subtext.ImportExport
 			bmlPost.DateCreated = entry.DateCreated;
 			bmlPost.DateModified = entry.DateModified;
             bmlPost.PostType = (entry.PostType == PostType.Story) ? BlogPostTypes.Article : BlogPostTypes.Normal;
-            bmlPost.Views = (uint) 0; // I think we have this statistic in the db... right?
+            bmlPost.Views = 0; // I think we have this statistic in the db... right?
 		    
             if (entry.HasEntryName)
             {
